+++ jest.config.js
@@ -2,7 +2,7 @@ module.exports = {
 	preset: 'ts-jest',
 	testEnvironment: 'node',
 	rootDir: '.',
-	testMatch: ['<rootDir>/test/**/*.test.ts'],
+	testMatch: ['<rootDir>/tests/**/*.test.ts'],
 	moduleFileExtensions: ['ts', 'js'],
 	reporters: ['default', 'jest-junit'],
 	collectCoverage: true,