+++ src/utils/common.ts
@@ -65,6 +65,20 @@ export function isQuoted(val: string): boolean {
 	return /^["'](.*)["']$/.test(val);
 }
 
+export function isMultilineInterpolation(val: string): boolean {
+	return /^`[\s\S]*`$/m.test(val) && val.includes('\n');
+}
+
+/**
+ * Encloses code in brackets and possibly spaces
+ *
+ * @param bracketSpacing Specifies whether or not to insert spaces before and after the code
+ * @param code Code that is enclosed in brackets
+ */
+export function handleBracketSpacing(bracketSpacing: boolean, code: string): string {
+	return bracketSpacing ? `{{ ${code} }}` : `{{${code}}}`;
+}
+
 // Copy of https://github.com/prettier/prettier/blob/master/src/common/util.js#L647
 export function makeString(
 	rawContent: string,
@@ -72,19 +86,19 @@ export function makeString(
 	unescapeUnnecessaryEscapes: boolean = false
 ): string {
 	const otherQuote = enclosingQuote === '"' ? "'" : '"';
-	const newContent = rawContent.replace(/\\([\s\S])|(['"])/g, (match, escaped, quote) => {
+	const newContent = rawContent.replace(/\\([\s\S])|(['"])/g, (match, escaped: "'" | '"', quote: "'" | '"') => {
 		if (escaped === otherQuote) {
 			return escaped;
 		}
 		if (quote === enclosingQuote) {
-			return '\\' + quote;
+			return `\\${quote}`;
 		}
 		if (quote) {
 			return quote;
 		}
 		return unescapeUnnecessaryEscapes && /^[^\\nrvtbfux\r\n\u2028\u2029"'0-7]$/.test(escaped)
 			? escaped
-			: '\\' + escaped;
+			: `\\${escaped}`;
 	});
 	return enclosingQuote + newContent + enclosingQuote;
 }