+++ tests/embed/vue/formatted.vue
@@ -1,0 +1,9 @@
+<template lang="pug">
+.text(color="primary", disabled)
+</template>
+
+<script>
+export default {
+  name: "Hello",
+};
+</script>